requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.1.0
phonenumberslite==8.13.11
sqlparse==0.4.4
tzdata==2023.3
